api/hum_service/pyproject.toml
[tool.poetry]
name = "hum_service"
version = "0.1.0"
description = ""
authors = ["Reddo team"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.75.1"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
python-multipart = "^0.0.5"
python-magic = "^0.4.25"
faiss-cpu = "^1.7.2"
numpy = "^1.22.3"
scipy = "1.4.1"
torch = "1.10.0"
torchvision = "0.11.1"
librosa = "0.8.1"
pydub = "^0.25.1"
tqdm = "4.64.0"
pandas = "1.3.5"
ffmpeg = "^1.4"
boto3 = "^1.22.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
python-magic-bin = "^0.4.14"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
